split weather tests into current and forecast
